Remove nodes that only contain a single token.  Instead, allow tokens directly in these positions as well as nodes.
This has two benefits.  First, it eliminates a large amount of allocations.  Previously, nearly all identifiers,
literals, primitive types, etc., would lead to double allocations.  An allocation for the token and an allocation
for the node wrapper.  Second, it results in a speedup as well.  Partly because there is less garbage (and less
collections) and partly because constructing some nodes are actually expensive.  For example, when instantiating
an IdentifierNameSyntax, the following constructors would run:

IdentifierNameSyntax -> NameSyntax -> TypeSyntax -> UnaryExpressionSyntax -> ExpressionSyntax -> SyntaxNode

These six constructors did add up when constructing millions of nodes.  Now, the token can be used as is
with no extra CPU or memory cost.