Removing parameters from pullTypeCheck.
The second parameter ("reportDiagnostics") was always set to true.  So it was unnecessary.
The first parameter ("refresh") was either set to true, or was set to false right after
the compiler was created.  In that latter case, we ended up running the same code as if
'refresh' were true.  As such, both parameters were unnecessary and could be removed.
This will hopefully simplify and clarify the underlying compiler API.  Right now you just
need to call 'pullTypeCheck' if any files have been added/removed from the compiler.
Ideally, in the future, the client won't need to be aware of this and the compiler will
do it itself.