Don't compute if we need an 'extends' clause in the AST translator.  It breaks in incremental scenarios.
For example, if we reuse a ModuleDeclaration entirely, then we won't traverse into it to discover
what's inside.  That means we'll miss a potential ClassDeclaration contained within and won't
set the bit properly.  Instead, just compute this in the emitter.  It only involves a top
level walk of modules, so it should be fast in pretty much all cases.