Properly bind a generic type when it is aliased through a module alias.
Note: my preference would actually be to allow a type alias to be generic itself.
However, this turned out to be problematic as sharing the generic type's type
parameters was unfeasiable.  For now, this seems like the easier and safer fix,
and is more in line with how the rest of the resolver works.