Moving code related to file paths out of the parser.
We had some code that checked if a relative path was used in an import statement
in a module declaration.  This check is fine.  It just shouldn't exist at the syntax
layer (syntax does not care about things like file paths). I've moved this to
the type check portion of the code, and i've also updated the error ID.