Report errors when continue/break statements are used improperly.
continue statements must be within an iteration statement.
continue statements with a target must reference a target of an iteration statement.

break statements without a target must be in an iteration or switch statement.
break statements with a target must reference an enclosing labeled statement.

break/contiue targets cannot cross a function boundary.
labels cannot use the same name as an enclosing label.