Do not pass hte 'incontextualassignment' flag when resolving a logical-or operation.
Note: this affected the 'noImplicitAny' feature.  It used to always report an error
if it saw an empty object literal that wasn't in a contextual assignment.  That meant
that we were now reporting an error for the following pattern:

var v: string[] = something || []

In this case though it is totally bogus to report a no-implicit-any error.  The type
of [] is "undefined[]" and as that never gets widened to "any[]" there was no implicit
"any".  I've fixed this by only reporting this error if we have an non-any array that was
actually widened to an "any[]".