Introduce the concept of a root decl.
A root decl can store all the information shared by it and
any of its child declaratoins (for example, the filename it
was generated from).  This allows us to save space on all
our decls (at the cost of logarythmic access to that information
from any decl).  In practice this has no slowdown for us.

The end goal is to allow us to store more information at the
decl level without having to bloat all decls.