Remove global semantic info chains and binders.
Now, the compiler itself owns the single semantic
info chain for the system.  The chain itself has
a mapping from file name to binders.

I've also added a link from root decls to the semantic
chain.  With this link it's now possible for a decl to
bind itself to a symbol as it can get the appropriate
binder from the chain that it points to.