Move the emit portion of the compiler to a proper API.
Prior to this change, emitting was an act of telling the actual compiler to generate
the emit output and actually write it to the io host.  This imperative style approach
to emitting is not appropriate for a compiler that we want to actually want to expose
as an analysis system.

Now, instead of having the compiler do the actual work of producing the result and
doing IO, the compiler is only responsible for producing the results (which is
essentially a set of (<file name> + <emitted content>) tuples).  The consumer of the
compiler API (the LS, the test harnesses, the batch compiler, etc.) can then decide
what they want to do with that result.  Now, outside of bugs, the emit API should not
be able to throw exceptions like it used to be able to do.