Improve performance of parser by using a switch to handle statements instead of chained if/elses.
While the previous system was clearer and more consistent with the rest of the
parser, we get a few percent improvement in speed with the switch since this is
such a hot spot in the compiler.  THe new code is still nearly as clear, so the
change seems worth it.