Update grammar for constructor declarations.
We now allow any call signature for a constructor.  In this was, if the user
writes:

    public constructor()

Or

    constructor<X>()

Or

    constructor(): number

We will consider it a constructor (and not a member function).  We will, of course,
also provide appropriate messages stating that the constructor isn't legal.

If someone wants an actual member called 'constructor' they can use a string literal
as hte property name.